refactor PollableValue
